Move muc domain group functions to their own file
